@@ -24,9 +24,9 @@ copyright = '2019, The TF Encrypted Authors'
 author = 'The TF Encrypted Authors'
 
 # The short X.Y version
-version = '0.8.0'
+version = '0.9.0'
 # The full version, including alpha/beta/rc tags
-release = '0.8.0'
+release = '0.9.0'
 
 
 # -- General configuration ---------------------------------------------------
