@@ -2,11 +2,8 @@
 
 These use TensorFlow's native dtypes tf.int32 and tf.int64 for the given float
 encoding being used (fixed-point, etc.)."""
-from __future__ import absolute_import
 
-import abc
 import math
-from typing import Dict
 from typing import List
 from typing import Optional
 from typing import Tuple
@@ -19,7 +16,6 @@ from ..operations import aux
 from ..operations import secure_random
 from .factory import AbstractConstant
 from .factory import AbstractFactory
-from .factory import AbstractPlaceholder
 from .factory import AbstractTensor
 from .factory import AbstractVariable
 from .helpers import inverse
@@ -31,53 +27,69 @@ from .shared import patches2im
 
 
 def native_factory(
-    NATIVE_TYPE, EXPLICIT_MODULUS=None,
+    NATIVE_TYPE,
+    EXPLICIT_MODULUS=None,
 ):  # pylint: disable=invalid-name
     """Constructs the native tensor Factory."""
 
     class Factory(AbstractFactory):
         """Native tensor factory."""
 
-        def tensor(self, value):
+        def tensor(self, initial_value):
 
-            if isinstance(value, tf.Tensor):
-                if value.dtype is not self.native_type:
-                    value = tf.cast(value, dtype=self.native_type)
-                return DenseTensor(value)
-
-            if isinstance(value, np.ndarray):
-                value = tf.convert_to_tensor(value, dtype=self.native_type)
-                return DenseTensor(value)
-
+            if isinstance(initial_value, Tensor):
+                return initial_value
+            elif isinstance(initial_value, tf.Tensor):
+                if initial_value.dtype is not self.native_type:
+                    initial_value = tf.cast(initial_value, dtype=self.native_type)
+                    return Tensor(initial_value)
+                else:
+                    return Tensor(initial_value)
+            elif isinstance(initial_value, np.ndarray):
+                initial_value = tf.convert_to_tensor(
+                    initial_value, dtype=self.native_type
+                )
+                return Tensor(initial_value)
             else:
-                # Give it a last try
-                value = np.array(value)
-                value = tf.convert_to_tensor(value, dtype=self.native_type)
-                return DenseTensor(value)
+                raise TypeError(
+                    "Don't know how to handle {}".format(type(initial_value))
+                )
 
-            raise TypeError("Don't know how to handle {}".format(type(value)))
-
-        def constant(self, value):
-
-            if isinstance(value, np.ndarray):
-                value = tf.constant(value, dtype=self.native_type)
-                return Constant(value)
-
-            raise TypeError("Don't know how to handle {}".format(type(value)))
+        def constant(self, initial_value):
+            if isinstance(initial_value, Tensor):
+                return initial_value
+            elif isinstance(initial_value, tf.Tensor):
+                if initial_value.dtype is not self.native_type:
+                    initial_value = tf.cast(initial_value, dtype=self.native_type)
+                    return Constant(initial_value)
+                else:
+                    return Constant(initial_value)
+            elif isinstance(initial_value, np.ndarray):
+                constant_value = tf.constant(initial_value, dtype=self.native_type)
+                return Constant(constant_value)
+            else:
+                raise TypeError(
+                    "Don't know how to handle {}".format(type(initial_value))
+                )
 
         def variable(self, initial_value):
 
-            if isinstance(initial_value, (tf.Tensor, np.ndarray)):
+            if isinstance(initial_value, tf.Variable):
                 return Variable(initial_value)
-
-            if isinstance(initial_value, Tensor):
-                return Variable(initial_value.value)
-
-            msg = "Don't know how to handle {}"
-            raise TypeError(msg.format(type(initial_value)))
-
-        def placeholder(self, shape):
-            return Placeholder(shape)
+            elif isinstance(initial_value, (tf.Tensor, np.ndarray)):
+                variable_value = tf.Variable(
+                    initial_value, dtype=self.native_type, trainable=False
+                )
+                return Variable(variable_value)
+            elif isinstance(initial_value, Tensor):
+                variable_value = tf.Variable(
+                    initial_value.value, dtype=self.native_type, trainable=False
+                )
+                return Variable(variable_value)
+            else:
+                raise TypeError(
+                    "Don't know how to handle {}".format(type(initial_value))
+                )
 
         @property
         def min(self):
@@ -113,19 +125,23 @@ def native_factory(
             maxval = self.max if maxval is None else maxval
 
             if secure_random.supports_seeded_randomness():
-                seed = secure_random.secure_seed()
-                return UniformTensor(
-                    shape=shape, seed=seed, minval=minval, maxval=maxval
+                value = secure_random.seeded_random_uniform(
+                    shape=shape,
+                    dtype=NATIVE_TYPE,
+                    minval=minval,
+                    maxval=maxval,
+                    seed=secure_random.secure_seed(),
                 )
+                return Tensor(value)
 
             if secure_random.supports_secure_randomness():
                 sampler = secure_random.random_uniform
             else:
-                sampler = tf.random_uniform
+                sampler = tf.random.uniform
             value = sampler(
                 shape=shape, minval=minval, maxval=maxval, dtype=NATIVE_TYPE
             )
-            return DenseTensor(value)
+            return Tensor(value)
 
         def sample_seeded_uniform(
             self,
@@ -149,7 +165,7 @@ def native_factory(
             maxval = self.max if maxval is None else maxval
 
             if secure_random.supports_seeded_randomness():
-                # Don't use UniformTensor for lazy sampling here, because the `seed`
+                # Don't use lazy sampling here, because the `seed`
                 # might be something (e.g., key) we want to protect, and we cannot
                 # send it to another party
                 value = secure_random.seeded_random_uniform(
@@ -159,27 +175,43 @@ def native_factory(
                     maxval=maxval,
                     seed=seed,
                 )
-                return DenseTensor(value)
+                return Tensor(value)
             else:
-                value = tf.random.stateless_uniform(
-                    shape, seed, minval=minval, maxval=maxval, dtype=NATIVE_TYPE
-                )
-                return DenseTensor(value)
+                if NATIVE_TYPE not in (tf.int32, tf.int64):
+                    value = tf.random.stateless_uniform(
+                        shape, seed, minval=minval, maxval=maxval, dtype=tf.int32
+                    )
+                    value = tf.cast(value, NATIVE_TYPE)
+                else:
+                    value = tf.random.stateless_uniform(
+                        shape,
+                        seed,
+                        minval=minval,
+                        maxval=maxval,
+                        dtype=NATIVE_TYPE,
+                    )
+                return Tensor(value)
 
         def sample_bounded(self, shape, bitlength: int):
-            maxval = 2 ** bitlength
+            maxval = 2**bitlength
             assert maxval <= self.max
 
             if secure_random.supports_seeded_randomness():
-                seed = secure_random.secure_seed()
-                return UniformTensor(shape=shape, seed=seed, minval=0, maxval=maxval)
+                value = secure_random.seeded_random_uniform(
+                    shape=shape,
+                    dtype=NATIVE_TYPE,
+                    minval=0,
+                    maxval=maxval,
+                    seed=secure_random.secure_seed(),
+                )
+                return Tensor(value)
 
             if secure_random.supports_secure_randomness():
                 sampler = secure_random.random_uniform
             else:
-                sampler = tf.random_uniform
+                sampler = tf.random.uniform
             value = sampler(shape=shape, minval=0, maxval=maxval, dtype=NATIVE_TYPE)
-            return DenseTensor(value)
+            return Tensor(value)
 
         def sample_bits(self, shape):
             return self.sample_bounded(shape, bitlength=1)
@@ -187,42 +219,29 @@ def native_factory(
         def stack(self, xs: list, axis: int = 0):
             assert all(isinstance(x, Tensor) for x in xs)
             value = tf.stack([x.value for x in xs], axis=axis)
-            return DenseTensor(value)
+            return Tensor(value)
 
         def concat(self, xs: list, axis: int):
             assert all(isinstance(x, Tensor) for x in xs)
             value = tf.concat([x.value for x in xs], axis=axis)
-            return DenseTensor(value)
+            return Tensor(value)
 
-        def where(self, condition, x, y, v2=False):
+        def where(self, condition, x, y):
             if not isinstance(condition, tf.Tensor):
                 msg = "Don't know how to handle `condition` of type {}"
                 raise TypeError(msg.format(type(condition)))
-            if not v2:
-                # Try to solve the broadcasting problem in a naive way.
-                # Not a comprehensive implementation.
-                if condition.shape != x.shape:
-                    shape = tf.broadcast_static_shape(
-                        tf.broadcast_static_shape(condition.shape, x.shape), y.shape
-                    )
-                    tile_shape = [
-                        (shape[i] // condition.shape[i]) for i in range(len(shape))
-                    ]
-                    condition = tf.tile(condition, tile_shape)
-                value = tf.where(condition, x.value, y.value)
-            else:
-                value = tf.compat.v2.where(condition, x.value, y.value)
-            return DenseTensor(value)
+            value = tf.where(condition, x.value, y.value)
+            return Tensor(value)
 
         def tile(self, x, multiples):
-            return DenseTensor(tf.tile(x.value, multiples))
+            return Tensor(tf.tile(x.value, multiples))
 
         def scatter_nd(self, indices, updates, shape):
-            return DenseTensor(tf.scatter_nd(indices, updates.value, shape))
+            return Tensor(tf.scatter_nd(indices, updates.value, shape))
 
-    FACTORY = Factory()  # pylint: disable=invalid-name
+    FACTORY = Factory()
 
-    def _lift(x, y) -> Tuple["Tensor", "Tensor"]:
+    def _lift(x, y) -> Tuple["Tensor", "Tensor"]:  # noqa:F821
 
         if isinstance(x, Tensor) and isinstance(y, Tensor):
             return x, y
@@ -242,19 +261,20 @@ def native_factory(
     class Tensor(AbstractTensor):
         """Base class for other native tensor classes."""
 
-        @property
-        @abc.abstractproperty
-        def value(self):
-            pass
+        def __init__(self, value: tf.Tensor):
+            self._value = value
 
         @property
-        @abc.abstractproperty
+        def value(self):
+            return self._value
+
+        @property
         def shape(self):
-            pass
+            return self._value.shape
 
         def identity(self):
             value = tf.identity(self.value)
-            return DenseTensor(value)
+            return Tensor(value)
 
         def to_native(self) -> tf.Tensor:
             return self.value
@@ -273,6 +293,10 @@ def native_factory(
         @property
         def factory(self):
             return FACTORY
+
+        @property
+        def device(self):
+            return self._value.device
 
         def __add__(self, other):
             x, y = _lift(self, other)
@@ -305,66 +329,68 @@ def native_factory(
             return self.mul(-1)
 
         def __getitem__(self, slc):
-            return DenseTensor(self.value[slc])
+            return Tensor(self.value[slc])
 
         def add(self, other):
             x, y = _lift(self, other)
             value = x.value + y.value
             if EXPLICIT_MODULUS is not None:
                 value %= EXPLICIT_MODULUS
-            return DenseTensor(value)
+            return Tensor(value)
 
         def sub(self, other):
             x, y = _lift(self, other)
             value = x.value - y.value
             if EXPLICIT_MODULUS is not None:
                 value %= EXPLICIT_MODULUS
-            return DenseTensor(value)
+            return Tensor(value)
 
         def mul(self, other):
             x, y = _lift(self, other)
             value = x.value * y.value
             if EXPLICIT_MODULUS is not None:
                 value %= EXPLICIT_MODULUS
-            return DenseTensor(value)
+            return Tensor(value)
 
         def matmul(self, other):
             x, y = _lift(self, other)
             value = tf.matmul(x.value, y.value)
             if EXPLICIT_MODULUS is not None:
                 value %= EXPLICIT_MODULUS
-            return DenseTensor(value)
+            return Tensor(value)
 
         def bit_gather(self, start, stride):
             value = aux.bit_gather(self.value, start, stride)
-            return DenseTensor(value)
+            return Tensor(value)
 
         def bit_split_and_gather(self, stride):
             value = aux.bit_split_and_gather(self.value, stride)
-            return DenseTensor(value)
+            return Tensor(value)
 
         def xor_indices(self):
             value = aux.xor_indices(self.value)
-            return DenseTensor(value)
+            return Tensor(value)
 
-        def im2col(self, h_filter, w_filter, stride, padding):
-            i2c = im2col(self.value, h_filter, w_filter, stride=stride, padding=padding)
-            return DenseTensor(i2c)
+        def im2col(self, h_filter, w_filter, strides, padding):
+            i2c = im2col(
+                self.value, h_filter, w_filter, strides=strides, padding=padding
+            )
+            return Tensor(i2c)
 
-        def im2patches(self, patch_size, stride, padding, data_format="NCHW"):
+        def im2patches(self, patch_size, strides, padding, data_format="NCHW"):
             i2p = im2patches(
                 self.value,
                 patch_size,
-                stride=stride,
+                strides=strides,
                 padding=padding,
                 data_format=data_format,
             )
-            return DenseTensor(i2p)
+            return Tensor(i2p)
 
         def patches2im(
             self,
             patch_size,
-            stride,
+            strides,
             padding,
             img_size=None,
             consolidation="SUM",
@@ -373,70 +399,70 @@ def native_factory(
             p2i = patches2im(
                 self.value,
                 patch_size,
-                stride=stride,
+                strides=strides,
                 padding=padding,
                 img_size=img_size,
                 consolidation=consolidation,
                 data_format=data_format,
             )
-            return DenseTensor(p2i)
+            return Tensor(p2i)
 
-        def conv2d(self, other, stride: int, padding: str = "SAME"):
+        def conv2d(self, other, strides: list, padding: str = "SAME"):
             if EXPLICIT_MODULUS is not None:
                 # TODO(Morten) any good reason this wasn't implemented for PrimeTensor?
                 raise NotImplementedError()
             x, y = _lift(self, other)
-            return DenseTensor(conv2d(x.value, y.value, stride=stride, padding=padding))
+            return Tensor(conv2d(x.value, y.value, strides=strides, padding=padding))
 
-        def batch_to_space_nd(self, block_shape, crops):
-            value = tf.batch_to_space_nd(self.value, block_shape, crops)
-            return DenseTensor(value)
+        def batch_to_space(self, block_shape, crops):
+            value = tf.batch_to_space(self.value, block_shape, crops)
+            return Tensor(value)
 
-        def space_to_batch_nd(self, block_shape, paddings):
-            value = tf.space_to_batch_nd(self.value, block_shape, paddings)
-            return DenseTensor(value)
+        def space_to_batch(self, block_shape, paddings):
+            value = tf.space_to_batch(self.value, block_shape, paddings)
+            return Tensor(value)
 
         def mod(self, k: int):
             value = self.value % k
             if EXPLICIT_MODULUS is not None:
                 value %= EXPLICIT_MODULUS
-            return DenseTensor(value)
+            return Tensor(value)
 
         def transpose(self, perm):
-            return DenseTensor(tf.transpose(self.value, perm))
+            return Tensor(tf.transpose(self.value, perm))
 
         def strided_slice(self, args, kwargs):
-            return DenseTensor(tf.strided_slice(self.value, *args, **kwargs))
+            return Tensor(tf.strided_slice(self.value, *args, **kwargs))
 
         def gather(self, indices: list, axis: int = 0):
-            return DenseTensor(tf.gather(self.value, indices, axis=axis))
+            return Tensor(tf.gather(self.value, indices, axis=axis))
 
         def split(self, num_split: Union[int, list], axis: int = 0):
             values = tf.split(self.value, num_split, axis=axis)
-            return [DenseTensor(value) for value in values]
+            return [Tensor(value) for value in values]
 
         def scatter_nd(self, indices, shape):
             value = tf.scatter_nd(indices, self.value, shape)
-            return DenseTensor(value)
+            return Tensor(value)
 
         def reverse(self, axis):
             value = tf.reverse(self.value, axis)
-            return DenseTensor(value)
+            return Tensor(value)
 
         def reshape(self, axes: Union[tf.Tensor, List[int]]):
-            return DenseTensor(tf.reshape(self.value, axes))
+            return Tensor(tf.reshape(self.value, axes))
 
         def negative(self):
             value = tf.negative(self.value)
             if EXPLICIT_MODULUS is not None:
                 value %= EXPLICIT_MODULUS
-            return DenseTensor(value)
+            return Tensor(value)
 
         def reduce_sum(self, axis, keepdims=None):
             value = tf.reduce_sum(self.value, axis, keepdims)
             if EXPLICIT_MODULUS is not None:
                 value %= EXPLICIT_MODULUS
-            return DenseTensor(value)
+            return Tensor(value)
 
         def cumsum(self, axis, exclusive, reverse):
             value = tf.cumsum(
@@ -444,11 +470,11 @@ def native_factory(
             )
             if EXPLICIT_MODULUS is not None:
                 value %= EXPLICIT_MODULUS
-            return DenseTensor(value)
+            return Tensor(value)
 
         def reduce_max(self, axis=None, keepdims=False):
             value = tf.reduce_max(self.value, axis, keepdims)
-            return DenseTensor(value)
+            return Tensor(value)
 
         def equal_zero(self, factory=None):
             factory = factory or FACTORY
@@ -466,18 +492,18 @@ def native_factory(
         def truncate(self, amount, base=2):
             if base == 2:
                 return self.right_shift(amount)
-            factor = base ** amount
+            factor = base**amount
             factor_inverse = inverse(factor, self.factory.modulus)
             return (self - (self % factor)) * factor_inverse
 
         def right_shift(self, bitlength):
-            return DenseTensor(tf.bitwise.right_shift(self.value, bitlength))
+            return Tensor(tf.bitwise.right_shift(self.value, bitlength))
 
         def expand_dims(self, axis: Optional[int] = None):
-            return DenseTensor(tf.expand_dims(self.value, axis))
+            return Tensor(tf.expand_dims(self.value, axis))
 
         def squeeze(self, axis: Optional[List[int]] = None):
-            return DenseTensor(tf.squeeze(self.value, axis=axis))
+            return Tensor(tf.squeeze(self.value, axis=axis))
 
         def cast(self, factory):
             return factory.tensor(self.value)
@@ -488,7 +514,7 @@ def native_factory(
         def bitwise_or(self, other):
             x, y = _lift(self, other)
             value = tf.bitwise.bitwise_or(x.value, y.value)
-            return DenseTensor(value)
+            return Tensor(value)
 
         def __xor__(self, other):
             return self.bitwise_xor(other)
@@ -496,7 +522,7 @@ def native_factory(
         def bitwise_xor(self, other):
             x, y = _lift(self, other)
             value = tf.bitwise.bitwise_xor(x.value, y.value)
-            return DenseTensor(value)
+            return Tensor(value)
 
         def __and__(self, other):
             return self.bitwise_and(other)
@@ -504,26 +530,26 @@ def native_factory(
         def bitwise_and(self, other):
             x, y = _lift(self, other)
             value = tf.bitwise.bitwise_and(x.value, y.value)
-            return DenseTensor(value)
+            return Tensor(value)
 
         def __invert__(self):
             return self.invert()
 
         def invert(self):
             value = tf.bitwise.invert(self.value)
-            return DenseTensor(value)
+            return Tensor(value)
 
         def __lshift__(self, bitlength):
             return self.left_shift(bitlength)
 
         def left_shift(self, bitlength):
-            return DenseTensor(tf.bitwise.left_shift(self.value, bitlength))
+            return Tensor(tf.bitwise.left_shift(self.value, bitlength))
 
         def __rshift__(self, bitlength):
             """
-      Arithmetic shift.
-      Please refer to `self.logical_rshift` for a logical right shift.
-      """
+            Arithmetic shift.
+            Please refer to `self.logical_rshift` for a logical right shift.
+            """
             return self.right_shift(bitlength)
 
         def logical_rshift(self, bitlength):
@@ -547,107 +573,43 @@ def native_factory(
             mask = ~((-1) << (total - bitlength))
             x = tf.bitwise.right_shift(self.value, bitlength)
             x = tf.bitwise.bitwise_and(x, mask)
-            return DenseTensor(x)
+            return Tensor(x)
 
         def bit_reverse(self):
             value = aux.bit_reverse(self.value)
-            return DenseTensor(value)
+            return Tensor(value)
 
-    class DenseTensor(Tensor):
-        """Public native Tensor class."""
-
-        def __init__(self, value):
-            self._value = value
-
-        @property
-        def shape(self):
-            return self._value.shape
-
-        @property
-        def value(self):
-            return self._value
-
-        @property
-        def support(self):
-            return [self._value]
-
-    class UniformTensor(Tensor):
-        """Class representing a uniform-random, lazily sampled tensor.
-
-    Lazy sampling optimizes communication by sending seeds in place of
-    fully-expanded tensors."""
-
-        def __init__(self, shape, seed, minval, maxval):
-            self._seed = seed
-            self._shape = shape
-            self._minval = minval
-            self._maxval = maxval
-
-        @property
-        def shape(self):
-            return self._shape
-
-        @property
-        def value(self):
-            with tf.name_scope("expand-seed"):
-                return secure_random.seeded_random_uniform(
-                    shape=self._shape,
-                    dtype=NATIVE_TYPE,
-                    minval=self._minval,
-                    maxval=self._maxval,
-                    seed=self._seed,
-                )
-
-        @property
-        def support(self):
-            return [self._seed]
-
-    class Constant(DenseTensor, AbstractConstant):
+    class Constant(Tensor, AbstractConstant):
         """Native Constant class."""
 
-        def __init__(self, constant: tf.Tensor) -> None:
-            assert isinstance(constant, tf.Tensor)
-            super(Constant, self).__init__(constant)
+        def __init__(self, constant_value: tf.Tensor) -> None:
+            super(Constant, self).__init__(constant_value)
 
         def __repr__(self) -> str:
             return "Constant(shape={})".format(self.shape)
 
-    class Placeholder(DenseTensor, AbstractPlaceholder):
-        """Native Placeholder class."""
-
-        def __init__(self, shape: List[int]) -> None:
-            self.placeholder = tf.placeholder(NATIVE_TYPE, shape=shape)
-            super(Placeholder, self).__init__(self.placeholder)
-
-        def __repr__(self) -> str:
-            return "Placeholder(shape={})".format(self.shape)
-
-        def feed(self, value: np.ndarray) -> Dict[tf.Tensor, np.ndarray]:
-            assert isinstance(value, np.ndarray), type(value)
-            return {self.placeholder: value}
-
-    class Variable(DenseTensor, AbstractVariable):
+    class Variable(Tensor, AbstractVariable):
         """Native Variable class."""
 
-        def __init__(self, initial_value: Union[tf.Tensor, np.ndarray]) -> None:
-            self.variable = tf.Variable(
-                initial_value, dtype=NATIVE_TYPE, trainable=False
-            )
-            self.initializer = self.variable.initializer
-            super(Variable, self).__init__(self.variable.read_value())
+        def __init__(self, variable_value: tf.Variable) -> None:
+            # TODO(zjn) need a better implementation to update Tensor's `_value`
+            # after variable has been assigned
+            # also for case bool factory and int100 factory
+            super(Variable, self).__init__(variable_value.read_value())
+            self.variable = variable_value
 
         def __repr__(self) -> str:
             return "Variable(shape={})".format(self.shape)
 
-        def assign_from_native(self, value: np.ndarray) -> tf.Operation:
-            assert isinstance(value, np.ndarray), type(value)
-            return self.assign_from_same(FACTORY.tensor(value))
+        def assign(self, value: Union[Tensor, np.ndarray]) -> None:
+            if isinstance(value, Tensor):
+                return self.variable.assign(value.value)
+            elif isinstance(value, np.ndarray):
+                return self.variable.assign(value)
+            else:
+                raise TypeError("Don't know how to handle {}".format(type(value)))
 
-        def assign_from_same(self, value: Tensor) -> tf.Operation:
-            assert isinstance(value, Tensor), type(value)
-            return tf.assign(self.variable, value.value).op
-
-        def read_value(self):
-            return DenseTensor(self.variable.read_value())
+        def read_value(self) -> Tensor:
+            return Tensor(self.variable.read_value())
 
     return FACTORY
